+++ packages/midway-decorator/src/rpc/hsf.ts
@@ -1,0 +1,17 @@
+import { ScopeEnum, saveClassMetadata, saveModule, HSF_KEY } from '../common';
+import { Scope } from '../annotation';
+
+export interface HSFOpts {
+  interfaceName?: string;
+  version?: string;
+  group?: string;
+  namespace?: string;
+}
+
+export function HSF(hsfOption: HSFOpts = {}): ClassDecorator {
+  return (target: any) => {
+    saveModule(HSF_KEY, target);
+    saveClassMetadata(HSF_KEY, hsfOption, target);
+    Scope(ScopeEnum.Request)(target);
+  };
+}