+++ CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](https://github.com/midwayjs/midway/compare/v2.0.1...v2.0.2) (2020-03-13)
+
+
+### Bug Fixes
+
+* Add export hsf ([#422](https://github.com/midwayjs/midway/issues/422)) ([55f1e43](https://github.com/midwayjs/midway/commit/55f1e43fbb9bd442939a6bb504aa721297eaf631))
+
+
+
+
+
+## [2.0.1](https://github.com/midwayjs/midway/compare/v2.0.0...v2.0.1) (2020-03-13)
+
+
+### Features
+
+* add hsf decorator ([#421](https://github.com/midwayjs/midway/issues/421)) ([d5afed3](https://github.com/midwayjs/midway/commit/d5afed3ace4e3570b29a2c789b2683f0cd4fd697))
+
+
+
+
+
+# [2.0.0](https://github.com/midwayjs/midway/compare/v2.0.0-beta.16...v2.0.0) (2020-03-13)
+
+**Note:** Version bump only for package midway
+
+
+
+
+
+# [2.0.0-beta.16](https://github.com/midwayjs/midway/compare/v2.0.0-beta.15...v2.0.0-beta.16) (2020-03-12)
+
+
+### Bug Fixes
+
+* 修复循环引用bug ([#419](https://github.com/midwayjs/midway/issues/419)) ([8852c6c](https://github.com/midwayjs/midway/commit/8852c6c55de8975aea3df2978bf50425378379e6))
+
+
+
+
+
 # [2.0.0-beta.15](https://github.com/midwayjs/midway/compare/v2.0.0-beta.14...v2.0.0-beta.15) (2020-03-06)