+++ packages/midway-web/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](https://github.com/midwayjs/midway/compare/v2.0.1...v2.0.2) (2020-03-13)
+
+**Note:** Version bump only for package midway-web
+
+
+
+
+
+## [2.0.1](https://github.com/midwayjs/midway/compare/v2.0.0...v2.0.1) (2020-03-13)
+
+**Note:** Version bump only for package midway-web
+
+
+
+
+
+# [2.0.0](https://github.com/midwayjs/midway/compare/v2.0.0-beta.16...v2.0.0) (2020-03-13)
+
+**Note:** Version bump only for package midway-web
+
+
+
+
+
+# [2.0.0-beta.16](https://github.com/midwayjs/midway/compare/v2.0.0-beta.15...v2.0.0-beta.16) (2020-03-12)
+
+**Note:** Version bump only for package midway-web
+
+
+
+
+
 # [2.0.0-beta.15](https://github.com/midwayjs/midway/compare/v2.0.0-beta.14...v2.0.0-beta.15) (2020-03-06)
 
 **Note:** Version bump only for package midway-web